Corrigindo formato de temperatura referente

+++ src/bme280_sensor.c
@@ -128,7 +128,7 @@ int8_t user_i2c_write(uint8_t reg_addr, const uint8_t *data, uint32_t len, void 
  * @retval BME280_E_NVM_COPY_FAILED - Error: NVM copy failed
  *
  */
-int8_t stream_sensor_data_forced_mode(struct bme280_dev *dev);
+float stream_sensor_data_forced_mode(struct bme280_dev *dev);
 
 /*!
  * @brief This function starts execution of the program.
@@ -261,7 +261,7 @@ void print_sensor_data(struct bme280_data *comp_data)
 /*!
  * @brief This API reads the sensor temperature, pressure and humidity data in forced mode.
  */
-int8_t stream_sensor_data_forced_mode(struct bme280_dev *dev)
+float stream_sensor_data_forced_mode(struct bme280_dev *dev)
 {
     /* Variable to define the result */
     int8_t rslt = BME280_OK;
